Prisma/Code/Src/Python/Prisma-dumy-generator-AAA/requirements.txt
opencv-python
pillow
pytesseract
numpy
unidecode
pdf2image
jsonschema
requests
tqdm
python-docx
reportlab
Faker
